+++ Aoc09/Program.cs
@@ -14,7 +14,7 @@ if (args.Length == 0) {
     var output = new CollectOutput();
     var quineVm = new IntVM("109,1,204,-1,1001,100,1,100,1008,100,16,101,1006,101,0,99",
             null, output);
-    var code = (int[])quineVm.Memory.Clone();
+    var code = (long[])quineVm.Memory.Clone();
     quineVm.Run();
     AssertEqualLists(output.Output, code, "Quine");
 } else {
