+++ IntCode/IO/ChainIO.cs
@@ -1,21 +1,21 @@
 namespace IntCode.IO;
 
 public sealed class ChainIO : VMBlockingInput, VMOutput {
-    public Queue<int> Values { get; } = new Queue<int>();
+    public Queue<long> Values { get; } = new Queue<long>();
 
-    public void Feed(int value) {
+    public void Feed(long value) {
         Values.Enqueue(value);
     }
 
     public bool CanRead() => Values.Any();
 
-    public int Read() => Values.Dequeue();
+    public long Read() => Values.Dequeue();
 
     public void Reset() {
         Values.Clear();
     }
 
-    public void Write(int number) => Feed(number);
+    public void Write(long number) => Feed(number);
 
     public override string ToString() {
         if (Values.Any()) {
